requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
spotipy>=2.22.0
numpy>=1.21.0
python-dotenv>=1.0.0
requests>=2.28.0